fix(kfx): satisfy source admission boundaries
