+++ .buildchain/kfd/kfd-1/contract-world.witness.json
@@ -35,7 +35,7 @@
   "sourceVerificationRequired": false,
   "canonicalPolicy": {
     "path": "framework/contract/kungfu-agent-first-canonical-policy.json",
-    "sha256": "73a6ad82d689cbb4f2884ff063d8f774b8422d3c5630f57ef76a750b8c40eafa"
+    "sha256": "dd4b272b37e30f46e35ced4d6feb083af4aae4ca0c332c6b4b53f1a121bf7fbe"
   },
   "registry": {
     "path": "framework/contract/kungfu-contracts.registry.json",
@@ -53,9 +53,9 @@
     {
       "name": "kungfu-kfx",
       "sourcePath": "framework/kfx/kungfu-kfx.contract.json",
-      "sourceSha256": "0eb5b6ccfa8cabb4d9e9186e021ac189c07ce42ee5302b611216272b289cec81",
+      "sourceSha256": "de1fed32395786194d38961ca222a70a4fa8d71bfe4bcf848d4fb35a69607ca7",
       "artifactPath": "config/kungfu-kfx.contract.json",
-      "expectedSha256": "0eb5b6ccfa8cabb4d9e9186e021ac189c07ce42ee5302b611216272b289cec81",
+      "expectedSha256": "de1fed32395786194d38961ca222a70a4fa8d71bfe4bcf848d4fb35a69607ca7",
       "byteForByte": true
     },
     {
