+++ developer/sdk/kfd/kfd-1/release-gate.json
@@ -88,13 +88,13 @@
       "standard": "KFD-1",
       "standardKey": "kfd-1",
       "result": "passed",
-      "preBuildWitnessSha256": "3e132c02c0c7e9b68d2e768d4936698eeec927f2f11a2b3fc2beb55425edbdea",
+      "preBuildWitnessSha256": "1b76e7299e4ff2fd7b7e81ba246f0190fc0bfc13efc8ec992c5936258c306b94",
       "sourceHashes": {
-        "sha256": "4314e87aeaf2f9caa36246665663a36e613cee1ab44f087975569033c4dd0599",
+        "sha256": "c7bbf8a78affc2a5dfff75c31f457d5dceffdd81be2844feb12b8eaf25d76de5",
         "surfaceCount": 21
       },
       "artifactHashes": {
-        "sha256": "4ee9e3e6b1b8b8ca4622b00d7ac59939b3f5a9bd2d1d91fe769f47ad802fd4a5",
+        "sha256": "f8de40eb335f5f275410fbaad9901a18b9a0b91b1e86a6b13866f742110e813c",
         "surfaceCount": 21
       },
       "witness": {
@@ -134,7 +134,7 @@
         "sourceVerificationRequired": true,
         "canonicalPolicy": {
           "path": "framework/contract/kungfu-agent-first-canonical-policy.json",
-          "sha256": "73a6ad82d689cbb4f2884ff063d8f774b8422d3c5630f57ef76a750b8c40eafa"
+          "sha256": "dd4b272b37e30f46e35ced4d6feb083af4aae4ca0c332c6b4b53f1a121bf7fbe"
         },
         "registry": {
           "path": "framework/contract/kungfu-contracts.registry.json",
@@ -152,9 +152,9 @@
           {
             "name": "kungfu-kfx",
             "sourcePath": "framework/kfx/kungfu-kfx.contract.json",
-            "sourceSha256": "0eb5b6ccfa8cabb4d9e9186e021ac189c07ce42ee5302b611216272b289cec81",
+            "sourceSha256": "de1fed32395786194d38961ca222a70a4fa8d71bfe4bcf848d4fb35a69607ca7",
             "artifactPath": "config/kungfu-kfx.contract.json",
-            "expectedSha256": "0eb5b6ccfa8cabb4d9e9186e021ac189c07ce42ee5302b611216272b289cec81",
+            "expectedSha256": "de1fed32395786194d38961ca222a70a4fa8d71bfe4bcf848d4fb35a69607ca7",
             "byteForByte": true
           },
           {
@@ -397,8 +397,8 @@
           {
             "name": "kungfu-kfx",
             "sourcePath": "framework/kfx/kungfu-kfx.contract.json",
-            "expectedSha256": "0eb5b6ccfa8cabb4d9e9186e021ac189c07ce42ee5302b611216272b289cec81",
-            "actualSha256": "0eb5b6ccfa8cabb4d9e9186e021ac189c07ce42ee5302b611216272b289cec81",
+            "expectedSha256": "de1fed32395786194d38961ca222a70a4fa8d71bfe4bcf848d4fb35a69607ca7",
+            "actualSha256": "de1fed32395786194d38961ca222a70a4fa8d71bfe4bcf848d4fb35a69607ca7",
             "status": "passed",
             "reason": ""
           },
@@ -574,10 +574,10 @@
           {
             "name": "kungfu-kfx",
             "sourcePath": "framework/kfx/kungfu-kfx.contract.json",
-            "sourceSha256": "0eb5b6ccfa8cabb4d9e9186e021ac189c07ce42ee5302b611216272b289cec81",
+            "sourceSha256": "de1fed32395786194d38961ca222a70a4fa8d71bfe4bcf848d4fb35a69607ca7",
             "artifactPath": "config/kungfu-kfx.contract.json",
-            "expectedSha256": "0eb5b6ccfa8cabb4d9e9186e021ac189c07ce42ee5302b611216272b289cec81",
-            "actualSha256": "0eb5b6ccfa8cabb4d9e9186e021ac189c07ce42ee5302b611216272b289cec81",
+            "expectedSha256": "de1fed32395786194d38961ca222a70a4fa8d71bfe4bcf848d4fb35a69607ca7",
+            "actualSha256": "de1fed32395786194d38961ca222a70a4fa8d71bfe4bcf848d4fb35a69607ca7",
             "byteForByte": true,
             "status": "passed",
             "reason": ""
