+++ docs/adr/KF-ADR-019f86da-4f90-73f3-b027-c343b2bc8bcc.md
@@ -4,7 +4,7 @@ doc_type: architecture-decision
 adr_id: KF-ADR-019f86da-4f90-73f3-b027-c343b2bc8bcc
 decision_status: accepted
 implementation_status: partial
-implementation_prs: [https://github.com/kungfu-systems/kungfu/pull/568, https://github.com/kungfu-systems/kungfu/pull/734, https://github.com/kungfu-systems/kungfu/pull/815, https://github.com/kungfu-systems/kungfu/pull/906, https://github.com/kungfu-systems/kungfu/pull/922, https://github.com/kungfu-systems/kungfu/pull/1704, https://github.com/kungfu-systems/kungfu/pull/1718]
+implementation_prs: [https://github.com/kungfu-systems/kungfu/pull/568, https://github.com/kungfu-systems/kungfu/pull/734, https://github.com/kungfu-systems/kungfu/pull/815, https://github.com/kungfu-systems/kungfu/pull/906, https://github.com/kungfu-systems/kungfu/pull/922, https://github.com/kungfu-systems/kungfu/pull/1704, https://github.com/kungfu-systems/kungfu/pull/1718, https://github.com/kungfu-systems/kungfu/pull/1728]
 review_state: self-reviewed
 sensitivity: public
 sources: [local-files, user-decision]
@@ -76,6 +76,15 @@ before side effects; explicit recovery Warrants authorize disable/remove
 without trusting package code, and terminal receipts settle the resulting
 Work, Warrant, Episode, Settlement, and Cut. Reference-aware collection,
 runtime grant enforcement, and terminal cutover remain later stages, so
+implementation remains partial.
+
+PR #1728 binds admitted lifecycle state to an explicit runtime capability
+grant. The accepted snapshot retains the exact declaration, policy, approval,
+Passport, Work/Warrant, Cut, revision, generation, and confinement roots; host
+launch recomputes and matches them before execution. Package, policy,
+capability, generation, or replay drift fails closed, and Product System
+metadata does not alter execution eligibility. Reference-aware collection,
+recursive Control Suite dogfood, and terminal cutover remain later stages, so
 implementation remains partial.
 
 ## Decision
