+++ framework/api/src/capability/index.ts
@@ -14,7 +14,6 @@ export * from './work-loop.js';
 export * from './workspace.js';
 export * from './storage.js';
 export * from './kfx-host.js';
-export * from './kfx-control.js';
 export * from './exit.js';
 export * from './query.js';
 export * from './profile.js';
